+++ crates/nu-parser/Cargo.toml
@@ -19,6 +19,8 @@ num-bigint = { version = "0.3.1", features = ["serde"] }
 num-traits = "0.2.14"
 serde = "1.0.118"
 shellexpand = "2.1.0"
+itertools = "0.10.0"
+smart-default = "0.6.0"
 
 dunce = "1.0.1"
 nu-errors = { version = "0.25.2", path = "../nu-errors" }
