+++ crates/nu-parser/src/lex/mod.rs
@@ -1,0 +1,6 @@
+pub mod lexer;
+mod token_group;
+pub mod tokens;
+
+#[cfg(test)]
+mod tests;
