+++ crates/nu-parser/src/lib.rs
@@ -9,7 +9,8 @@ mod scope;
 mod shapes;
 mod signature;
 
-pub use lex::{block, lex, LiteBlock, LiteCommand, LiteGroup, LitePipeline};
+pub use lex::lexer::{block, lex};
+pub use lex::tokens::{LiteBlock, LiteCommand, LiteGroup, LitePipeline};
 pub use parse::{classify_block, garbage, parse, parse_full_column_path, parse_math_expression};
 pub use path::expand_ndots;
 pub use scope::ParserScope;
